requirements.txt
-i https://pypi.org/simple
aioredis==2.0.1
aiosqlite==0.17.0; python_version >= '3.6'
anyio==3.7.0; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
asyncpg==0.27.0
bcrypt==4.0.1
click==8.1.3; python_version >= '3.7'
fastapi==0.97.0
fastapi-limiter==0.1.5
h11==0.14.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
iso8601==1.1.0; python_version < '4.0' and python_full_version >= '3.6.2'
passlib==1.7.4
pydantic==1.10.9; python_version >= '3.7'
pyjwt==2.7.0
pypika-tortoise==0.1.6; python_version >= '3.7' and python_version < '4.0'
python-multipart==0.0.6
pytz==2023.3
redis==4.5.5
sniffio==1.3.0; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
tortoise-orm==0.19.3
typing-extensions==4.6.3; python_version >= '3.7'
uvicorn==0.22.0
